WinDRBD: New function op_is_known
Newer drbd-utils query the netlink family of the Linux
kernel for known commands. We don't implement a full
netlink replacement but still want to query the kernel
for commands it knows. Therefore we introduced this ioctl
which checks if a certain command is known.

For example the command DRBD_ADM_COMMAND_PATHS is implemented
in DRBD 9.1 and DRBD 9.2 but not in DRBD 9.0. This DRBD
utils work with all DRBD versions (including older ones)
in a sane way that is without printing an error message
and printing the paths if the kernel module supports it.