DRBDmon: Fix incorrect object deallocation if saving the configuration fails
When changing a configuration entry using the command line arguments,
a changed configuration entry's allocation was only released after
attempting to save the configuration file.
If saving the configuration file fails, the corresponding method
throws an exception that triggers automatic deallocation of the
configuration entry that had already been linked into the
configuration. This can cause a memory access error during the
deallocation of the configuration when DBRDmon exits.
This fix moves the release statement into the correct position and
adds comments on how to use the set_entry method correctly.