drbdsetup,84: fix udev command
The new drbdsetup udev command always returned a non-zero exit code. This was
caused by returning "1" when no "*genl_info" was available. However, this is
expected, as the command function is called once for the resource, and another
time afterwards with empty genl_info to indicate completion.

Fix this by returning 0 for the case the genl_info is null. All the validation
is already handled by the outer layers.

Fixes: 18f80de3d9ff ("drbdsetup: add udev command")