drbdsetup: DRBD 8 CLI syntax compatibility mode
CLI syntax differs slightly between DRBD 8 and 9, which leads to
breakage when upgrading DRBD major versions.
The "usual path", using a configuration file and drbdadm commands, is
already handled correctly.
However, if a user invokes drbdsetup directly and without drbdadm (for
example, in a script), this usage would break when upgrading to a DRBD 9
kernel module.

We plan to ship DRBD 9 in the Linux kernel at some point, so prepare for
this scenario and try to keep these old usage patterns compatible.

Specificially, the "new-resource" and "connect" commands are affected.
"new-resource" now takes a node-id which it did not before, and
"connect" now takes two node-ids, where it took the nodes' IP addresses
directly before.

If we detect an old usage of one of these commands, do some trickery in
the command line parsing logic to mimick the old behavior.

For "new-resource", convince the kernel module to leave the node id
uninitialized for now.
For "connect", issue an extra "new-peer" and "new-path" command,
creating a peer with the specified IP address. The node ids are
arbitrated using the specified IP addresses. The "lower" IP gets node id
0, the other one gets 1.
Then, send the actual "connect" command with these generated node ids.