build,rpm: don't try to rm drbdmon.8.gz in preun
This is most likely a fallout from changes introduced in
f962315c30e901258bd151f33994d85f32905cee

We have "versioned man pages" and create symlinks to them (e.g.,
drbdmeta.8.gz -> drbdmeta-9.0.8.gz). Creating and deleting these is
managed in post and preun scripts. drbdmon is special as there is no
versioned man page, there is only drbdmon.8.gz (the final file, no
symlinks).

As is, we tried to rm the file twice, once in preun and then as part of
"files". This resulted in a completely harmless message that a
non-existent file was tried to get removed.