drbdadm: reset ctx in ctx_by_* functions
Source code pattern is:
	r = ctx_by_name(&ctx, "id", ...);
	if (!ctx.res)
		log error
	if (r)
		exit 1

If you specify a number of resources on the command line,
and the first one is defined, we forgot to log the error,
but still do the exit 1.

We need to reset ctx with each call to ctx_by_...().

This also changes the logic:
we no longer exit on the first "non-existing" resource, but log the error,
accumulate the "exit 1", then still continue with all other resource names
that are listed on the command line.