wrap_printf: fix wrong intendation after newline
If the current column position is already at or beyond the wrap limit,
our home-grown "pretty print with indentation" wrapper adds spurious
indentation after a newline if you call it with an indentation != 0.

That skews the indentation of the next line
if it is supposed to be at a different indentation level.

For example if the last device statistic line happend to end at or beyond col 80,
the first connection line would align with those stats, and not with column 2:
```
      bm-writes:35790 upper-pending:0 lower-pending:0 al-suspended:no blocked:no
      connection-name node-id:0 connection:Connected role:Secondary tls:no
```

We could fix all calls to `wrap_printf(..., "\n")` to `(0, "\n")`.
I chose to try and fix it generically in wrap_printf() itself.

Still not correct for all possible "format strings"
or argument values, but better than before.