Revert "drbdmeta: fix return/exit with negative values"
This reverts commit 3d5e36589c6915508083c3766c8ecfb2e95dc820.

Commit 3d5e36589c69 mapped all return codes other than 0 to 1. The
various sub-commands return various ad-hoc return codes. Some of them
-1 for specific fatal errors.

The POSIX documentation for exit says:

  The value of status may be 0, EXIT_SUCCESS, EXIT_FAILURE, or any
  other value, though only the least significant 8 bits (that is,
  status & 0377) shall be available from wait() and waitpid(); the
  full value shall be available from waitid() and in the siginfo_t
  passed to a signal handler for SIGCHLD.

So, that -1 gets mapped to 255.

It has value for users to be able to determine the details of a
failure by checking for a specific exit value.

So, it happened that an exit value of
20  ... indicates a DRBD device is configured, although it should not be
255 ... indicates metadata is not valid

These are undocumented, as no one wants to take on the burden of
sanitizing the existing exit codes and documenting them.
This commit restores the status quo as it was and as some DRBD-8.4
users rely on it.