adjust,v9: Improve adjust for diskless primaries
When drbd.conf changes how a diskless node reaches nodes holding the
data, applying the changes might be nontrivial.

Consider a diskless primary, having one peer with an UpToDate
disk. One might remove the connection between the diskless node and
that node with the backing disk and add another connection to another
diskful node at the same time. That is a valid change when the two
diskful nodes have a connection.

So far, `drbdadm adjust` has failed to apply such a change, as it
attempts to remove the connection first and then add the other one.

Fixing that by leveraging an explicit dependency from the del-peer to
a wait-connect for the new peer.

The failure of a disconnect command triggered the previous
implementation, so it relied on the DRBD state engine itself to reject
the first disconnect. The new implementation is overly cautious, so it
attempts to establish the new connection before deleting the old one,
even if it is not strictly necessary.