drbdadm: fix sh-list-adjustable command
It appears that phil forgot to add sh-list-adjustable to the is_adjust
condition, which led to a segfault and connection-mesh errors.

I am also adding a test case to highlight the issue. That unveiled
that if there were a running resource that needs to be taken down, the
sh-list-adjustable would output that resource name twice. Also fixing
that corner case.