WinDRBD: Flush registry keys after install/remove bus device.
When the system is reset a few seconds after the WinDRBD
installer is run we end up in a state where Windows complains
that the WinDRBD driver is not installed and there is also
no bus device. With that patch the registry entries should
be made persistent and WinDRBD should be installed normally
even if the machine is reset right after the installer was
run. Note that there is no way to fix it if the system is
reset *while* the installer is running (installer is non-atomic
as any other installer on Windows too).