drbdadm create-md: do not spuriously re-generate device uuid
We may just now convert from 8.4 to 9 style meta data (or back).
If we try to read-dev-uuid before the conversion, that will fail,
and we would generate a new device-uuid.

Instead, create-md first, possibly converting the format.
Only then try to read-dev-uuid.

While at it, if we already have a device uuid,
there is no need to re-write it.