drbdadm create-md: do not reset effective-size on format conversion
During format conversion, keep as much as possible of what we know
about the device state.

A regression as a side effect of introducing additional optional
command line parameter --effective-size to the create-md sub-command
with drbd-utils version 9.28 resulted in casually resetting
the effective size to 0 (or whatever was given on the command line)
when converting between meta data versions.

The next attach then looks like a "first" attach to the DRBD module.

When downgrading from 9 to 8.4 style meta data,
our "initialize the bitmap during first attach" approach
effectively caused a spurious full sync.

Upgrading 8.4 to 9 did the same, but with 9 we take a
"trust the on disk bitmap anyways" approach,
even during an "apparently first" attach,
which effectively hid the potential problem.