drbdsetup: Bring the node-id heuristics in sync
I failed to notice earlier that we have a node-id heuristic in
drbdadm_postparse.c when we added the second implementation, now in
drbdsetup_compat84.c.

That is important so that a resource you bring up using drbdsetup in
compatibility mode generates the same node IDs as if you bring it up
using drbdadm, and let drbdadm perform the node ID heuristic.