tnf-drbd-fence.py: Consider that the fence history might be empty
So far, the code has assumed that the XML output from stonith_admin
includes a 'last-fenced' element. Well, that is not true if the fence
history is empty.
That caused an instance to exit with an exception, which uncovered
incomplete exception handling in parallel running instances waiting on
the lockfile.

Fix both aspects.